Moved UserRegisterForm to User.kt
