+++ src/main/kotlin/com/bpavuk/models/User.kt
@@ -13,6 +13,14 @@ data class User(
 @Serializable
 data class UserLoginForm(val username: String, val password: String)
 
+@Serializable
+data class UserRegisterForm(
+    val username: String,
+    val password: String,
+    val imageUrl: String
+)
+
+
 object Users : Table() {
     val id = integer("id").autoIncrement()
     val username = varchar("username", 45)
