--- src/main/kotlin/com/bpavuk/models/UserRegisterForm.kt
@@ -1,10 +1,0 @@
-package com.bpavuk.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserRegisterForm(
-    val username: String,
-    val password: String,
-    val imageUrl: String
-)
